Pirathaban/dev

+++ src/main/java/com/ii/testautomation/controllers/ProjectController.java
@@ -5,7 +5,9 @@ import com.ii.testautomation.dto.search.ProjectSearch;
 import com.ii.testautomation.enums.RequestStatus;
 import com.ii.testautomation.response.common.BaseResponse;
 import com.ii.testautomation.response.common.ContentResponse;
+import com.ii.testautomation.response.common.FileResponse;
 import com.ii.testautomation.response.common.PaginatedContentResponse;
+import com.ii.testautomation.service.ModulesService;
 import com.ii.testautomation.service.ProjectService;
 import com.ii.testautomation.utils.Constants;
 import com.ii.testautomation.utils.EndpointURI;
@@ -19,13 +21,18 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @CrossOrigin
 public class ProjectController {
     @Autowired
     private ProjectService projectService;
+    @Autowired
+    private ModulesService modulesService;
     @Autowired
     private StatusCodeBundle statusCodeBundle;
 
@@ -46,49 +53,62 @@ public class ProjectController {
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getSaveProjectSuccessMessage()));
-
     }
-//    @PostMapping(value = EndpointURI.PROJECT_IMPORT)
-//    public ResponseEntity<Object> saveProjectByImportFile(MultipartFile multipartFile)
-//    {
-//        List<ProjectRequest> projectRequestList=projectService.importProjectFile(multipartFile);
-//        for (ProjectRequest projectRequest:projectRequestList
-//             ) {
-//            if(!Utils.isNotNullAndEmpty(projectRequest.getName()))
-//            {
-//                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
-//                        statusCodeBundle.getProjectFileEmptyCode(),
-//                        statusCodeBundle.getProjectNameEmptyMessage()));
-//            }
-//            if(Utils.isNotNullAndEmpty(projectRequest.getCode()))
-//            {
-//                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
-//                        statusCodeBundle.getProjectFileEmptyCode(),
-//                        statusCodeBundle.getProjectCodeEmptyMessage()));
-//            }
-//            if(Utils.isNotNullAndEmpty(projectRequest.getDescription()))
-//            {
-//                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
-//                        statusCodeBundle.getProjectFileEmptyCode(),
-//                        statusCodeBundle.getProjectDescriptionEmptyMessage()));
-//            }
-//            if (projectService.existByProjectName(projectRequest.getName())) {
-//                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
-//                        statusCodeBundle.getProjectAlReadyExistCode(),
-//                        statusCodeBundle.getProjectNameAlReadyExistMessage()));
-//            }
-//            if (projectService.existByProjectCode(projectRequest.getCode())) {
-//                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
-//                        statusCodeBundle.getProjectAlReadyExistCode(),
-//                        statusCodeBundle.getProjectCodeAlReadyExistMessage()));
-//            }
-//        }
-//        projectService.saveProjectList(projectRequestList);
-//        return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
-//                statusCodeBundle.getCommonSuccessCode(),
-//                statusCodeBundle.getSaveProjectSuccessMessage()));
-//
-//    }
+
+    @PostMapping(value = EndpointURI.PROJECT_IMPORT)
+    public ResponseEntity<Object> importProjectFile(@RequestParam MultipartFile multipartFile) {
+        Map<String, List<Integer>> errorMessages = new HashMap<>();
+        List<ProjectRequest> projectRequestList;
+        try {
+            if (multipartFile.getOriginalFilename().endsWith(".csv")) {
+                projectRequestList = projectService.csvToProjectRequest(multipartFile.getInputStream());
+            } else if (projectService.hasExcelFormat(multipartFile)) {
+                projectRequestList = projectService.excelToProjectRequest(multipartFile);
+            } else {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getFileFailureCode(), statusCodeBundle.getFileFailureMessage()));
+            }
+
+            for (int rowIndex = 2; rowIndex <= projectRequestList.size() + 1; rowIndex++) {
+                ProjectRequest projectRequest = projectRequestList.get(rowIndex - 2);
+
+                if (!Utils.isNotNullAndEmpty(projectRequest.getName())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameEmptyMessage(), rowIndex);
+                }
+                if (!Utils.isNotNullAndEmpty(projectRequest.getCode())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectCodeEmptyMessage(), rowIndex);
+                }
+                if (!Utils.isNotNullAndEmpty(projectRequest.getDescription())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectDescriptionEmptyMessage(), rowIndex);
+                }
+                if (projectService.existByProjectName(projectRequest.getName())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameAlReadyExistMessage(), rowIndex);
+                }
+                if (projectService.existByProjectCode(projectRequest.getCode())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectCodeAlReadyExistMessage(), rowIndex);
+                }
+            }
+
+            if (!errorMessages.isEmpty()) {
+                return ResponseEntity.ok(new FileResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getFailureCode(),
+                        statusCodeBundle.getProjectFileImportValidationMessage(),
+                        errorMessages));
+            } else {
+                for (ProjectRequest projectRequest : projectRequestList) {
+                    projectService.saveProject(projectRequest);
+                }
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
+                        statusCodeBundle.getCommonSuccessCode(),
+                        statusCodeBundle.getSaveProjectSuccessMessage()));
+            }
+        } catch (IOException e) {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getFailureCode(),
+                    statusCodeBundle.getSaveProjectValidationMessage()));
+        }
+    }
+
 
     @PutMapping(value = EndpointURI.PROJECT)
     public ResponseEntity<Object> editProject(@RequestBody ProjectRequest projectRequest) {
@@ -149,11 +169,19 @@ public class ProjectController {
                     statusCodeBundle.getProjectNotExistCode(),
                     statusCodeBundle.getProjectNotExistsMessage()));
         }
+        if (modulesService.existsModuleByProjectId(id)) {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getProjectIdDependentCode(),
+                    statusCodeBundle.getProjectIdDependentMessage()));
+        }
+
         projectService.deleteProject(id);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(), statusCodeBundle.getDeleteProjectSuccessMessage()
         ));
     }
-
-
 }
+
+
+
+
