+++ src/main/java/com/ii/testautomation/controllers/SubModulesController.java
@@ -1,27 +1,33 @@
 package com.ii.testautomation.controllers;
 
 import com.ii.testautomation.dto.request.SubModulesRequest;
-import com.ii.testautomation.dto.response.SubModulesResponse;
-import com.ii.testautomation.dto.search.ProjectSearch;
 import com.ii.testautomation.dto.search.SubModuleSearch;
 import com.ii.testautomation.enums.RequestStatus;
 import com.ii.testautomation.response.common.BaseResponse;
 import com.ii.testautomation.response.common.ContentResponse;
+import com.ii.testautomation.response.common.FileResponse;
 import com.ii.testautomation.response.common.PaginatedContentResponse;
 import com.ii.testautomation.service.MainModulesService;
 import com.ii.testautomation.service.SubModulesService;
+import com.ii.testautomation.service.TestCasesService;
 import com.ii.testautomation.utils.Constants;
 import com.ii.testautomation.utils.EndpointURI;
 import com.ii.testautomation.utils.StatusCodeBundle;
-import org.apache.coyote.Response;
+import com.ii.testautomation.utils.Utils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @CrossOrigin
@@ -30,27 +36,24 @@ public class SubModulesController {
     private SubModulesService subModulesService;
     @Autowired
     private MainModulesService mainModulesService;
-
+    @Autowired
+    private TestCasesService testCasesService;
     @Autowired
     private StatusCodeBundle statusCodeBundle;
 
     @PostMapping(value = EndpointURI.SUBMODULE)
-    public ResponseEntity<Object> saveSubModules(@RequestBody SubModulesRequest subModulesRequest)
-    {
-        if(subModulesService.existsBySubModulesName(subModulesRequest.getName()))
-        {
+    public ResponseEntity<Object> saveSubModules(@RequestBody SubModulesRequest subModulesRequest) {
+        if (subModulesService.existsBySubModulesName(subModulesRequest.getName())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModuleNameAlReadyExistMessage()));
         }
-        if(subModulesService.existsBySubModulesPrefix(subModulesRequest.getPrefix()))
-        {
+        if (subModulesService.existsBySubModulesPrefix(subModulesRequest.getPrefix())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModulePrefixAlReadyExistMessage()));
         }
-        if (!mainModulesService.existsByMainModuleId(subModulesRequest.getMain_module_Id()))
-        {
+        if (!mainModulesService.isExistMainModulesId(subModulesRequest.getMain_module_Id())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getMainModulesNotExistCode(),
                     statusCodeBundle.getMainModuleNotExistsMessage()));
@@ -59,80 +62,120 @@ public class SubModulesController {
         return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getSaveSubModuleSuccessMessage()));
-
     }
-    @PutMapping(value=EndpointURI.SUBMODULE)
-    public ResponseEntity<Object> editSubModules(@RequestBody SubModulesRequest subModulesRequest)
-    {
-        if (!subModulesService.existsBySubModuleId(subModulesRequest.getId()))
-        {
+
+    @PostMapping(value = EndpointURI.SUBMODULE_IMPORT)
+    public ResponseEntity<Object> importSubModuleFile(@RequestParam MultipartFile multipartFile) {
+        Map<String, List<Integer>> errorMessages = new HashMap<>();
+        List<SubModulesRequest> subModulesRequestList = new ArrayList<>();
+        try (InputStream inputStream = multipartFile.getInputStream()) {
+            if (multipartFile.getOriginalFilename().endsWith(".csv")) {
+                subModulesRequestList = subModulesService.csvToSubModuleRequest(inputStream);
+            } else if (subModulesService.hasExcelFormat(multipartFile)) {
+                subModulesRequestList = subModulesService.excelToSubModuleRequest(multipartFile);
+            } else {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getFileFailureCode(), statusCodeBundle.getFileFailureMessage()));
+            }
+            for (int rowIndex = 2; rowIndex <= subModulesRequestList.size() + 1; rowIndex++) {
+                SubModulesRequest subModulesRequest = subModulesRequestList.get(rowIndex - 2);
+
+                if (!Utils.isNotNullAndEmpty(subModulesRequest.getName())) {
+                    subModulesService.addToErrorMessages(errorMessages, statusCodeBundle.getSubModuleNameEmptyMessage(), rowIndex);
+                }
+                if (!Utils.isNotNullAndEmpty(subModulesRequest.getPrefix())) {
+                    subModulesService.addToErrorMessages(errorMessages, statusCodeBundle.getSubModulePrefixEmptyMessage(), rowIndex);
+                }
+                if (subModulesService.existsBySubModulesPrefix(subModulesRequest.getPrefix())) {
+                    subModulesService.addToErrorMessages(errorMessages, statusCodeBundle.getSubModulePrefixAlReadyExistMessage(), rowIndex);
+                }
+                if (subModulesService.existsBySubModulesName(subModulesRequest.getName())) {
+                    subModulesService.addToErrorMessages(errorMessages, statusCodeBundle.getSubModuleNameAlReadyExistMessage(), rowIndex);
+                }
+                if (!mainModulesService.isExistMainModulesId(subModulesRequest.getMain_module_Id())) {
+                    subModulesService.addToErrorMessages(errorMessages, statusCodeBundle.getMainModuleNotExistsMessage(), rowIndex);
+                }
+            }
+            if (!errorMessages.isEmpty()) {
+                return ResponseEntity.ok(new FileResponse(RequestStatus.FAILURE.getStatus(),
+                        statusCodeBundle.getFailureCode(),
+                        statusCodeBundle.getSubModuleFileImportValidationMessage(),
+                        errorMessages));
+            } else {
+                for (SubModulesRequest subModulesRequest : subModulesRequestList) {
+                    subModulesService.saveSubModules(subModulesRequest);
+                }
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
+                        statusCodeBundle.getCommonSuccessCode(),
+                        statusCodeBundle.getSaveSubModuleSuccessMessage()));
+            }
+        } catch (IOException e) {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getFailureCode(),
+                    statusCodeBundle.getSaveProjectValidationMessage()));
+        }
+    }
+
+    @PutMapping(value = EndpointURI.SUBMODULE)
+    public ResponseEntity<Object> editSubModules(@RequestBody SubModulesRequest subModulesRequest) {
+        if (!subModulesService.existsBySubModuleId(subModulesRequest.getId())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesNotExistCode(),
                     statusCodeBundle.getSubModuleNotExistsMessage()));
         }
-        if (!mainModulesService.existsByMainModuleId(subModulesRequest.getMain_module_Id()))
-        {
+        if (!mainModulesService.isExistMainModulesId(subModulesRequest.getMain_module_Id())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getMainModulesNotExistCode(),
                     statusCodeBundle.getMainModuleNotExistsMessage()));
         }
-        if(subModulesService.isUpdateSubModuleNameExits(subModulesRequest.getName(),subModulesRequest.getId()))
-        {
+        if (subModulesService.isUpdateSubModuleNameExits(subModulesRequest.getName(), subModulesRequest.getId())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModuleNameAlReadyExistMessage()));
         }
-        if(subModulesService.isUpdateSubModulePrefixExits(subModulesRequest.getPrefix(),subModulesRequest.getId()))
-        {
+        if (subModulesService.isUpdateSubModulePrefixExits(subModulesRequest.getPrefix(), subModulesRequest.getId())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModulePrefixAlReadyExistMessage()));
         }
-
         subModulesService.saveSubModules(subModulesRequest);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getUpdateSubModuleSuccessMessage()));
-
     }
+
     @GetMapping(value = EndpointURI.SUBMODULE_BY_ID)
-    public ResponseEntity<Object> getSubModuleById(@PathVariable Long id)
-    {
-        if (!subModulesService.existsBySubModuleId(id))
-        {
+    public ResponseEntity<Object> getSubModuleById(@PathVariable Long id) {
+        if (!subModulesService.existsBySubModuleId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesNotExistCode(),
                     statusCodeBundle.getSubModuleNotExistsMessage()));
         }
-        return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULE,subModulesService.getSubModuleById(id),
+        return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULE, subModulesService.getSubModuleById(id),
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getGetSubModulesSuccessMessage()));
     }
 
     @GetMapping(value = EndpointURI.SUBMODULE_BY_MAIN_MODULE_ID)
-    public ResponseEntity<Object> getSubModuleByMainModuleId(@PathVariable Long id)
-    {
-        if (!mainModulesService.existsByMainModuleId(id))
-        {
+    public ResponseEntity<Object> getSubModuleByMainModuleId(@PathVariable Long id) {
+        if (!mainModulesService.isExistMainModulesId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getMainModulesNotExistCode(),
                     statusCodeBundle.getMainModuleNotExistsMessage()));
         }
-        List<SubModulesResponse> subModulesResponseList=subModulesService.getAllSubModuleByMainModuleId(id);
-        if(subModulesResponseList.isEmpty())
-        {
+        if (!subModulesService.existsByMainModuleId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getFailureCode(),
                     statusCodeBundle.getGetSubModuleNotHaveMainModuleId()));
         }
-
         return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULES,
                 subModulesService.getAllSubModuleByMainModuleId(id),
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getGetSubModulesSuccessMessage()));
     }
+
     @GetMapping(value = EndpointURI.SUBMODULES_SEARCH)
     public ResponseEntity<Object> getALlSubModuleWithMultiSearch(@RequestParam(name = "page") int page,
                                                                  @RequestParam(name = "size") int size,
@@ -141,24 +184,27 @@ public class SubModulesController {
                                                                  SubModuleSearch subModuleSearch) {
         Pageable pageable = PageRequest.of(page, size, Sort.Direction.valueOf(direction), sortField);
         PaginatedContentResponse.Pagination pagination = new PaginatedContentResponse.Pagination(page, size, 0, 0l);
-        return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULES, subModulesService.multiSearchSubModule(pageable,pagination,subModuleSearch),
+        return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULES, subModulesService.multiSearchSubModule(pageable, pagination, subModuleSearch),
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getGetAllSubModuleSuccessMessage()));
     }
+
     @DeleteMapping(value = EndpointURI.SUBMODULE_BY_ID)
-    public ResponseEntity<Object> deleteSubModuleById(@PathVariable Long id)
-    {
-        if (!subModulesService.existsBySubModuleId(id))
-        {
+    public ResponseEntity<Object> deleteSubModuleById(@PathVariable Long id) {
+        if (!subModulesService.existsBySubModuleId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesNotExistCode(),
                     statusCodeBundle.getSubModuleNotExistsMessage()));
+        }
+        if (testCasesService.existsBySubModuleId(id)) {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getSubModulesDependentCode(),
+                    statusCodeBundle.getSubModulesDependentMessage()));
         }
         subModulesService.deleteSubModuleById(id);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getDeleteSubModuleSuccessMessage()));
     }
-
 }
